Detach Head to attack player
When reaching the attack range of the lower half of the body, the head
detaches and continues pursuit of the player. Head will 'bite' when
reaching its attack range.

Right now, the Head will stay detached forever. Must make it so that it
reattaches after a while. Also, it seems that the Head can 'trap' the
player if it reaches it. Watch out.