Re-attach Head when losing Player
Finally started using two separate graphs for the AI. The head can
re-attach to the body after losing sight of the player. However, when
the Head detaches, the body doesn't fully get disabled--its 'Chase
Player' state is still active, which is an intended feature in Bolt.
Have to find a way to stop that state when disabling the lower half of
the monster.